backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic==0.42.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
sse-starlette==2.2.1
httpx==0.27.2

# Database & ORM
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Auth & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
